+++ .identity/03_github_environment_app_cd.tf
@@ -19,23 +19,7 @@ resource "github_repository_environment" "github_repository_environment_app_cd" 
   }
 }
 
-#tfsec:ignore:github-actions-no-plain-text-action-secrets # not real secret
-resource "github_actions_environment_secret" "azure_app_cd_tenant_id" {
-  repository      = var.github.repository
-  environment     = "${var.env}-app-cd"
-  secret_name     = "AZURE_TENANT_ID"
-  plaintext_value = data.azurerm_client_config.current.tenant_id
-}
-
-#tfsec:ignore:github-actions-no-plain-text-action-secrets # not real secret
-resource "github_actions_environment_secret" "azure_app_cd_subscription_id" {
-  repository      = var.github.repository
-  environment     = "${var.env}-app-cd"
-  secret_name     = "AZURE_SUBSCRIPTION_ID"
-  plaintext_value = data.azurerm_subscription.current.subscription_id
-}
-
-#tfsec:ignore:github-actions-no-plain-text-action-secrets # not real secret
+# TODO: delete
 resource "github_actions_environment_secret" "azure_app_cd_client_id" {
   repository      = var.github.repository
   environment     = "${var.env}-app-cd"