+++ apps/io-services-cms-webapp/src/webservice/controllers/__tests__/edit-service.test.ts
@@ -120,6 +120,10 @@ const mockBlobService = {
   createBlockBlobFromText: vi.fn((_, __, ___, cb) => cb(null, "any")),
 } as any;
 
+const mockServiceTopicDao = {
+  findAllNotDeletedTopics: vi.fn(() => TE.right(O.none)),
+} as any;
+
 afterEach(() => {
   vi.clearAllMocks();
 });
@@ -136,6 +140,7 @@ describe("editService", () => {
     subscriptionCIDRsModel,
     telemetryClient: mockAppinsights,
     blobService: mockBlobService,
+    serviceTopicDao: mockServiceTopicDao,
   });
 
   setAppContext(app, mockContext);